+++ c-api/resvg.h
@@ -17,9 +17,9 @@
 #include <stdint.h>
 
 #define RESVG_MAJOR_VERSION 0
-#define RESVG_MINOR_VERSION 30
+#define RESVG_MINOR_VERSION 31
 #define RESVG_PATCH_VERSION 0
-#define RESVG_VERSION "0.30.0"
+#define RESVG_VERSION "0.31.0"
 
 /**
  * @brief List of possible errors.